Add the capability to run tests inside DLL using a custom wrapper exe
When Catch2 tests are embedded in a DLL, they cannot be run by directly
running the output itself. Instead some kind of wrapper executable
must be used. The way such wrappers invoke the tests in the DLL
is a project-specific matter, so such a system requires that
the wrapper and its command line parameters are configurable.

This MR achieves this with two additional settings:
* `DllExecutor`, a path to the executable to run tests
  in DLL:s with.
* `DllExecutorCommandLine`, the command line parameters passed
  to the executor to run the tests inside some Source DLL.